chore: SOF V2 Conversion (#369)
# **Current PR Status**:

- Foundational setup is done. `WorldWithContext` is being used and I've
started our setup script in `EveTest.sol`
- When it comes time to deploy this for real, we can break the logic in
`EveTest` to a separate deploy script. I kept it in `EveTest` for now
for testing simplicity (no need to run and deploy a separate World and
run tests against it).
- I think I've added `scope` and `access` to all the appropriate places,
but this needs to be confirmed. Errors here will be apparent when
running tests.
- I've done a lot of work on implementing the permission matrix found
[here](https://miro.com/app/board/uXjVK_LSVxY=/). Still need to do one
last pass as @ccp-tezza updated it recently. I've been adding tests that
hit the access functions as part of the separate deployable tests.

## **Current Passing Tests**:
- `SmartStorageUnit`
- `InventoryInteract`
- `Inventory`
- `EphemeralInventory`
- `Deployable`
- `Fuel`
- `SmartTurret`
- `SmartGate`
- `Location`

## **My Process for Converting Tests**:
1. Inherit from `EveTest`. This will do all of the SOFv2 setup
2. Remove all old setup from the test. i.e. user addresses are handled,
`World` is now `WorldWithContext`
3. Make sure to create a new `classId` if needed and to instantiate the
smart object as that class. i.e. In DeployableTest we create a mock
deployable so we need to create a class id associated with this. Usually
this is handled in the top level create functions like
`CreateAndAchorSmartStorageUnit`
4. If there are `UnscopedSystemCall` errors, tag the class id with the
necessary system
5. If there are `Access` errors, prank as the right user in the test
when making the system calls

### Test Tips
- If you are hitting `EntityAlreadyExists` when instantiating, it is
most likely because test fuzzing is using a generated id that conflicts
with a hardcoded ID. This was happening a lot with `characterId`
conflicting with `smartObjectId`. You can add `vm.assume(smartObjectId
!= characterId)` to make sure this doesn't happen.
- When hitting `UnscopedSystemCall` you can search for the given
`systemId` in the codebase and the generated libs will pop up for easy
translation.

## **Final Confirmation**:
When all tests are passing, we should do a final pass and confirm:
- `Access` functions exactly match the permission matrix
- `context`, `scope`, and `access` are added to every system

---------

Co-authored-by: CCP Legolas <toniya@ccpgames.com>